adjust namespace for the settings classes
