+++ Core/JSONConverters/CameraConnectionConverter.cs
@@ -1,4 +1,4 @@
-﻿using RoverControlApp.MVVM.Model.Settings;
+﻿using RoverControlApp.Core.Settings;
 using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
