+++ MVVM/ViewModel/Grzyb_UIOverlay.cs
@@ -1,4 +1,5 @@
 using Godot;
+using RoverControlApp.Core;
 using RoverControlApp.MVVM.Model;
 using System.Collections.Generic;
 using System.Text;
